refactor: rename nameSpace.js to nameSpaceHelper.js
