+++ src/mods/nameSpaceHelper.js
