+++ src/mods/nameSpaceHelper.spec.js
